+++ src/main/java/com/georgster/game/plinko/PlinkoCommand.java
@@ -20,8 +20,6 @@ import discord4j.discordjson.json.ApplicationCommandRequest;
  */
 public class PlinkoCommand implements ParseableCommand {
 
-    private static final String PATTERN = "1|R 1|O";
-
     /**
      * {@inheritDoc}
      */
@@ -31,13 +29,13 @@ public class PlinkoCommand implements ParseableCommand {
 
         PlinkoGame game = new PlinkoGame(event); //Creates a PlinkoGame
 
-        subcommands.on(p -> {
+        subcommands.on(() -> {
             try {
                 game.startGame();
             } catch (IllegalStateException e) {
                 event.getGuildInteractionHandler().sendText(e.getMessage(), "Plinko");
             }
-        }, "play");
+        });
 
         subcommands.on(p -> {
             logger.append("- Showing a blank Plinko Board", LogDestination.NONAPI, LogDestination.API);
@@ -50,7 +48,7 @@ public class PlinkoCommand implements ParseableCommand {
      */
     @Override
     public PermissibleAction getRequiredPermission(List<String> args) {
-        if (args.contains("play")) {
+        if (args.isEmpty()) {
             return PermissibleAction.PLINKOGAME;
         } else {
             return PermissibleAction.DEFAULT;
@@ -62,7 +60,7 @@ public class PlinkoCommand implements ParseableCommand {
      */
     @Override
     public CommandParser getCommandParser() {
-        return new CommandParser(PATTERN);
+        return new CommandParser("1O");
     }
 
     /**
@@ -84,10 +82,7 @@ public class PlinkoCommand implements ParseableCommand {
                         .name("command")
                         .description("Play a game of Plinko!")
                         .type(ApplicationCommandOption.Type.STRING.getValue())
-                        .addChoice(ApplicationCommandOptionChoiceData.builder()
-                                .name("play")
-                                .value("play")
-                                .build())
+                        .required(false)
                         .addChoice(ApplicationCommandOptionChoiceData.builder()
                                 .name("board")
                                 .value("board")
@@ -101,7 +96,7 @@ public class PlinkoCommand implements ParseableCommand {
      */
     public String help() {
         return "Aliases: " + getAliases().toString() +
-        "\n- '!plinko play' to simulate a game of plinko" +
+        "\n- '!plinko' to simulate a game of plinko" +
         "\n- '!plinko board' to show an empty plinko board";
     }