+++ src/main/java/com/georgster/util/commands/exception/CommandParserException.java
@@ -1,16 +1,16 @@
 package com.georgster.util.commands.exception;
 
-import com.georgster.util.commands.NewCommandParser;
+import com.georgster.util.commands.CommandParser;
 
 /**
- * The super class for all {@link RuntimeException RuntimeExceptions} thrown by a {@link NewCommandParser}.
+ * The super class for all {@link RuntimeException RuntimeExceptions} thrown by a {@link CommandParser}.
  */
 public abstract class CommandParserException extends RuntimeException {
 
     /**
      * Creates a new {@link CommandParserException} with the provided message.
      * 
-     * @param message The message for why the {@link NewCommandParser} threw an exception.
+     * @param message The message for why the {@link CommandParser} threw an exception.
      */
     protected CommandParserException(String message) {
         super(message);